+++ src/lib.rs
@@ -86,6 +86,7 @@ pub use dynamic_graphql_derive::Interface;
 pub use dynamic_graphql_derive::Mutation;
 pub use dynamic_graphql_derive::MutationFields;
 pub use dynamic_graphql_derive::MutationRoot;
+pub use dynamic_graphql_derive::OneOfInput;
 #[doc = include_str!("./docs/resolved-object.md")]
 pub use dynamic_graphql_derive::ResolvedObject;
 #[doc = include_str!("./docs/resolved-object-fields.md")]